requirements.txt
asgiref==3.8.1
Django==4.2.11
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
drf-pretty-exception-handler==0.1.3
Faker==24.3.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
